Split out TranslatedResource & TranslatedResourceQuerySet
